requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alluka==0.3.3
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorlog==6.10.1
confspec==0.0.5
cryptography==46.0.3
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
hf-xet==1.2.0
hikari==2.5.0
hikari-lightbulb==3.2.2
hikari-miru==4.2.0
huggingface-hub==0.36.0
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
linkd==0.2.0
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pillow==12.1.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
python-dotenv==1.2.1
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.17.0
setuptools==80.9.0
SQLAlchemy==2.0.44
sympy==1.14.0
timm==1.0.24
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.6.3
uvloop==0.22.1
yarl==1.22.0
